requirements.txt
prometheus_client==0.5.0
python-neutronclient==6.11.0
python-ironicclient==2.5.1
python-dotenv==0.10.1
kubernetes==v7.0.1
configparser==3.7.1